+++ src/services/api.js
@@ -29,13 +29,12 @@ export async function fetchRestaurant({ restaurantId }) {
 }
 
 export async function postLogin({ email, password }) {
-  const url = 'https://eatgo-customer-api.ahastudio.com/session';
+  const url = 'https://eatgo-login-api.ahastudio.com/session';
   const response = await fetch(url, {
     method: 'POST',
     headers: {
       'Content-Type': 'application/json',
     },
-    mode: 'no-cors',
     body: JSON.stringify({ email, password }),
   });
